SYM-606: Disclose overflow in comparison changes
Moving preprocessing method into util repository file.
