+++ frontend/src/app/report/comparison-report.component.html
@@ -19,6 +19,18 @@
       [colors]="(legend | async)?.colorMap"
       unit="%"
     ></app-result-color-scale>
+    <app-pressure-chart *ngIf="report.chartDataPositive.links.length > 1"
+      [id]="'sankey-cmp-positive'" [diagramId]="'sankey-cmp-positive'"
+      [chartWeightThreshold]="getChartThresholdPercentage()"
+      [data]="formatChartData(report.chartDataPositive, metadata.bandDictionary)"
+      [locale]="locale"
+    ></app-pressure-chart>
+    <app-pressure-chart *ngIf="report.chartDataNegative.links.length > 1"
+      [id]="'sankey-cmp-negative'" [diagramId]="'sankey-cmp-negative'"
+      [chartWeightThreshold]="getChartThresholdPercentage()"
+      [data]="formatChartData(report.chartDataNegative, metadata.bandDictionary)"
+      [locale]="locale"
+    ></app-pressure-chart>
     <app-scenario-changes
       [name]="report.a.name"
       [scenarioChanges]="report.a.scenarioChanges"
@@ -35,18 +47,6 @@
       [comparisonReport]="true"
       [overflow]="report.b.overflow"
     ></app-scenario-changes>
-    <app-pressure-chart *ngIf="report.chartDataPositive.links.length > 1"
-      [id]="'sankey-cmp-positive'" [diagramId]="'sankey-cmp-positive'"
-      [chartWeightThreshold]="getChartThresholdPercentage()"
-      [data]="formatChartData(report.chartDataPositive, metadata.bandDictionary)"
-      [locale]="locale"
-    ></app-pressure-chart>
-    <app-pressure-chart *ngIf="report.chartDataNegative.links.length > 1"
-      [id]="'sankey-cmp-negative'" [diagramId]="'sankey-cmp-negative'"
-      [chartWeightThreshold]="getChartThresholdPercentage()"
-      [data]="formatChartData(report.chartDataNegative, metadata.bandDictionary)"
-      [locale]="locale"
-    ></app-pressure-chart>
     <app-impact-table
       class="pressure-impacts"
       [title]="'report.impact-table.pressure' | translate"
