+++ frontend/src/app/report/report.util.ts
@@ -1,7 +1,8 @@
 import { ChartData } from "@src/app/report/pressure-chart/pressure-chart.component";
-import { BandTypes, LayerData } from "@data/metadata/metadata.interfaces";
+import { BandType, BandTypes } from "@data/metadata/metadata.interfaces";
 import { cloneDeep } from "lodash";
 import { Report } from "@data/calculation/calculation.interfaces";
+import { ChangesProperty } from "@data/scenario/scenario.interfaces";
 
 export function relativeDifference(a: any, b: any) {
   const na = a as number,
@@ -52,4 +53,41 @@ export function bandTitlesWithOverflow(report: Report, bandDictionary: { [bandTy
   }
 
   return bandTitles;
+}
+
+export function setOverflowProperty(report: Report) {
+  let allChanges: { [key: number]: { [key: string]: ChangesProperty } } = {};
+  allChanges[0] = report.scenarioChanges.baseChanges;
+  allChanges = { ...allChanges, ...report.scenarioChanges.areaChanges };
+
+  // There's an argument to be made that this should be done in the backend.
+  // Opting for this solution after having tried the other route, since it's
+  // arguably (probably) a smaller and cheaper operation, to preprocess here.
+  // Also keep in mind that we'll be looking to implement a more thorough
+  // handling of the "overflow" phenomenon already in the subsequent release,
+  // where this part is very likely to be changed anyway. Making substantial
+  // changes that'd introduce complex behaviour to multiple entity classes and
+  // dto's in the backend service, which are likewise subject to change in the
+  // very short term, I feel is the less reasonable option at this point.
+
+  for (const changes of Object.entries(allChanges)) {
+    const index = +changes[0],
+      changeRef = index === 0 ?
+        report.scenarioChanges.baseChanges :
+        report.scenarioChanges.areaChanges[+changes[0]];
+    for (const change of Object.entries(changes[1])) {
+      const bandType = change[0];
+      for (const bandChange of Object.entries(change[1] as ChangesProperty)) {
+        const positiveChange =
+          (bandChange[1].offset && bandChange[1].offset > 0) ||
+          (bandChange[1].multiplier && bandChange[1].multiplier > 1) || false;
+        changeRef[bandType][bandChange[0]].hasOverflow =
+          positiveChange &&
+          report.overflow !== null &&
+          report.overflow[bandType as BandType].includes(+bandChange[0]);
+      }
+    }
+  }
+
+  return report;
 }
